test: (py) clean up utils/helpers split
"utils" are things test definition need
"helpers" are internal
